Conflicts:
	src/main/webapp/js/views/gestionCentrosSalud.html
	src/main/webapp/js/views/gestionUsuarios.html
	src/main/webapp/js/views/login.html
	src/main/webapp/js/views/vistaPaciente.html
	src/main/webapp/js/views/vistaSanitario.html

+++ src/main/java/uclm/grupo2/sigeva/http/UsuarioController.java
@@ -15,14 +15,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
+import uclm.grupo2.sigeva.dao.CitasDAO;
+import uclm.grupo2.sigeva.dao.TokenDAO;
 import uclm.grupo2.sigeva.dao.UsuarioDAO;
 import uclm.grupo2.sigeva.exceptions.CamposVaciosException;
 import uclm.grupo2.sigeva.exceptions.CredencialesInvalidasException;
+import uclm.grupo2.sigeva.exceptions.EsUnAdministradorException;
 import uclm.grupo2.sigeva.exceptions.FormatoDniException;
 import uclm.grupo2.sigeva.exceptions.FormatoPasswordException;
 import uclm.grupo2.sigeva.exceptions.NoEsTelefonoException;
+import uclm.grupo2.sigeva.exceptions.PacienteConCitasException;
+import uclm.grupo2.sigeva.exceptions.TokenBorradoException;
+import uclm.grupo2.sigeva.exceptions.UsuarioConVacunaException;
 import uclm.grupo2.sigeva.exceptions.UsuarioDuplicadoException;
 import uclm.grupo2.sigeva.exceptions.UsuarioInexistenteException;
+import uclm.grupo2.sigeva.model.CentroSalud;
+import uclm.grupo2.sigeva.model.Citas;
 import uclm.grupo2.sigeva.model.Usuario;
 
 @RestController
@@ -31,10 +39,17 @@ public class UsuarioController {
 
 	@Autowired
 	private UsuarioDAO user;
+	
+	@Autowired
+	private TokenDAO tokenLogin;
+	
+	@Autowired
+	private CitasDAO cita;
 		
 	@PostMapping("/insertUsers")
 	public String insertarUsuario(@RequestBody Usuario usuarios) {
 		try {
+			validarLogin();
 			Optional<Usuario> optUser = user.findByLogin(usuarios.getLogin());
 			if (optUser.isPresent())
 				throw new UsuarioDuplicadoException();
@@ -58,22 +73,26 @@ public class UsuarioController {
 		return "Usuario con id: "+usuarios.getId();
 	}
 	@GetMapping("/findAllUsers")
-	public List<Usuario> getUsuarios(){
+	public List<Usuario> getUsuarios() throws TokenBorradoException{
+		validarLogin();
 		return user.findAll();
-	}
-	@GetMapping("/findAllUsers/{id}")
-	public Optional<Usuario> getUsuario(@PathVariable String id){
-		return user.findById(id);
 	}
 	
 	@DeleteMapping("/deleteUser")
 	public String borrarUsuario(@RequestBody Usuario usuario) {
 		try {
+			validarLogin();
 			Optional<Usuario> optUser = user.findById(usuario.getId());
-			if (optUser.isPresent())
-				user.deleteById(usuario.getId());
-
-			else
+			if (optUser.isPresent()) {
+				if(!optUser.get().getRol().equals("Administrador")) {
+					if(optUser.get().getDosis()!=1) {
+						user.deleteById(usuario.getId());
+						cita.deleteAll(cita.getByPacienteOrderByNumCitaAsc(optUser.get()));
+					} else
+						throw new UsuarioConVacunaException();
+				} else 
+					throw new EsUnAdministradorException();
+			}else
 				throw new UsuarioInexistenteException();
 		} catch (Exception e) {
 			throw new ResponseStatusException(HttpStatus.NOT_FOUND, e.getMessage());
@@ -83,23 +102,21 @@ public class UsuarioController {
 	
 	@PostMapping("/modifyUser")
 	public String modificarUsuario(@RequestBody Usuario usuario) {
-		try {
-			
+		try {			
+			validarLogin();			
 			Optional<Usuario> optUser = user.findById(usuario.getId());
 			
 			if (optUser.isPresent()) {
+					List <Citas> cambiarUsuCitas = cita.getByPacienteOrderByNumCitaAsc(optUser.get());
 				 	Usuario preUsuario = optUser.get();
 				 	preUsuario.setPassword(usuario.getPassword());
+				 	comprobarCamposVacios(usuario);
 				 	preUsuario.setNombre(usuario.getNombre());
 				 	preUsuario.setApellidos(usuario.getApellidos());
 				 	preUsuario.setTelefono(usuario.getTelefono());
 				 	preUsuario.setDni(usuario.getDni());
 				 	preUsuario.setRol(usuario.getRol());
-				 	if(usuario.getCs()==null) {
-				 		preUsuario.setCs(preUsuario.getCs());
-					}else {
-				 	preUsuario.setCs(usuario.getCs());
-					}
+				 	preUsuario=cambiarCentro(usuario, preUsuario);
 					if(!validarMovil(preUsuario.getTelefono()))
 						throw new NoEsTelefonoException();
 					
@@ -112,7 +129,12 @@ public class UsuarioController {
 					if(!validarDni(preUsuario.getDni())) 
 						throw new FormatoDniException();
 					
-				 	user.save(preUsuario);			
+					for (int i=0; i<cambiarUsuCitas.size();i++) {
+						cambiarUsuCitas.get(i).setPaciente(preUsuario);
+						cita.save(cambiarUsuCitas.get(i));
+					}
+					
+				 	user.save(preUsuario);	
 			}
 			else
 				throw new UsuarioInexistenteException();
@@ -122,25 +144,19 @@ public class UsuarioController {
 		return "Usuario modificado";
 	}
 
-	@PostMapping("/iniciarSesion")
-	public Usuario iniciarSesion(@PathVariable String login, @PathVariable String password){
-		try {
-			Optional<Usuario> optUser = user.findByLogin(login);
-			if(optUser.isPresent()) {
-				Usuario usua = optUser.get();
-				if(DigestUtils.sha512Hex(password).equals(usua.getPassword())) {
-					return usua;
-				} else {
-					throw new CredencialesInvalidasException();
-				}
-			
-			}
-			return null;	
-		} catch(Exception e) {
-			throw new ResponseStatusException(HttpStatus.CONFLICT, e.getMessage());
+	private Usuario cambiarCentro(Usuario usuario, Usuario preUsuario) throws PacienteConCitasException {
+		if(usuario.getCs()==null) {
+	 		preUsuario.setCs(preUsuario.getCs());
+		}else {
+			if(cita.getByPacienteOrderByNumCitaAsc(preUsuario).isEmpty())
+				preUsuario.setCs(usuario.getCs());
+			else
+				throw new PacienteConCitasException();
 		}
+ 		return preUsuario;
+
+		
 	}
-	
 	private static boolean validarMovil(String telefono) {
 		if(telefono.length()!=9) {
 			return false;
@@ -191,4 +207,18 @@ public class UsuarioController {
 		}
 		return correcto;
 	}
+	
+	private void comprobarCamposVacios(Usuario usuario) throws CamposVaciosException {
+	if(usuario.getNombre().isEmpty() || usuario.getApellidos().isEmpty())
+ 		throw new CamposVaciosException();
+	}
+	private void validarLogin() throws TokenBorradoException {
+		if(tokenLogin.findAll().isEmpty())
+			throw new TokenBorradoException();
+
+    	List<Usuario> usuarios = user.getByLogin(tokenLogin.findAll().get(0).getLogin());
+    	Usuario usu = usuarios.get(0);
+        if(!usu.getRol().equals("Administrador"))
+            throw new TokenBorradoException();
+        }
 }
